requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
asyncpg==0.30.0
python-multipart==0.0.9
firebase-admin==6.4.0
google-cloud-pubsub==2.19.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
line-bot-sdk==3.21.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
